requirements.txt
pandas
numpy
seaborn
matplotlib
requests
python-decouple
